chore: review comments, int16->int32 conversion.

+++ tests/all_in_one.py
@@ -9,7 +9,7 @@ from time import perf_counter
 from typing import Optional
 
 import numpy as np
-from numba import cuda, int16, int32, int64, float32, float64
+from numba import cuda, int32, int32, int64, float32, float64
 from numba import from_dtype as numba_from_dtype
 from cubie.cuda_simsafe import activemask, all_sync, selp, compile_kwargs
 from cubie.cuda_simsafe import from_dtype as simsafe_dtype
@@ -1041,8 +1041,8 @@ def dirk_step_inline_factory(
             numba_precision[::1],
             numba_precision,
             numba_precision,
-            int16,
-            int16,
+            int32,
+            int32,
             numba_precision[::1],
             numba_precision[::1],
             int32[::1],
@@ -1169,14 +1169,14 @@ def dirk_step_inline_factory(
         #            Stage 0: may reuse cached values                     #
         # --------------------------------------------------------------- #
 
-        first_step = first_step_flag != int16(0)
+        first_step = first_step_flag != int32(0)
 
         # Only use cache if all threads in warp can - otherwise no gain
         use_cached_rhs = False
         if first_same_as_last and multistage:
             if not first_step:
                 mask = activemask()
-                all_threads_accepted = all_sync(mask, accepted_flag != int16(0))
+                all_threads_accepted = all_sync(mask, accepted_flag != int32(0))
                 use_cached_rhs = all_threads_accepted
         else:
             use_cached_rhs = False
@@ -1493,8 +1493,8 @@ def erk_step_inline_factory(
             numba_precision[::1],
             numba_precision,
             numba_precision,
-            int16,
-            int16,
+            int32,
+            int32,
             numba_precision[::1],
             numba_precision[::1],
             int32[::1],
@@ -1593,7 +1593,7 @@ def erk_step_inline_factory(
         if first_same_as_last and multistage:
             if not first_step_flag:
                 mask = activemask()
-                all_threads_accepted = all_sync(mask, accepted_flag != int16(0))
+                all_threads_accepted = all_sync(mask, accepted_flag != int32(0))
                 use_cached_rhs = all_threads_accepted
         else:
             use_cached_rhs = False
@@ -1876,8 +1876,8 @@ def firk_step_inline_factory(
             numba_precision[::1],
             numba_precision,
             numba_precision,
-            int16,
-            int16,
+            int32,
+            int32,
             numba_precision[::1],
             numba_precision[::1],
             int32[::1],
@@ -2201,8 +2201,8 @@ def rosenbrock_step_inline_factory(
             numba_precision[::1],
             numba_precision,
             numba_precision,
-            int16,
-            int16,
+            int32,
+            int32,
             numba_precision[::1],
             numba_precision[::1],
             int32[::1],
@@ -3263,8 +3263,8 @@ def loop_fn(initial_states, parameters, driver_coefficients, shared_scratch,
     controller_temp = persistent_local[local_controller_slice]
     step_persistent_local = persistent_local[local_step_slice]
 
-    first_step_flag = int16(1)
-    prev_step_accepted_flag = int16(1)
+    first_step_flag = int32(1)
+    prev_step_accepted_flag = int32(1)
 
     # ----------------------------------------------------------------------- #
     #                       Seed t=0 values                                   #
@@ -3383,7 +3383,7 @@ def loop_fn(initial_states, parameters, driver_coefficients, shared_scratch,
                 step_persistent_local,
                 proposed_counters,
             )
-            first_step_flag = int16(0)
+            first_step_flag = int32(0)
 
             niters = (step_status >> 16) & status_mask
             status |= step_status & status_mask
@@ -3439,8 +3439,8 @@ def loop_fn(initial_states, parameters, driver_coefficients, shared_scratch,
 
             prev_step_accepted_flag = selp(
                 accept,
-                int16(1),
-                int16(0),
+                int32(1),
+                int32(0),
             )
 
             # Predicated update of next_save; update if save is accepted.
